+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Unstyled table/th/td tags included in Send-HtmlMailMessage are now formatted
+
 ### Changed
 - LastLine property on Send-HtmlMailMessage no longer set by default